# exploration: speed up rescue_prime_regular benchmark (not for merging)