--- a/twenty-first/src/shared_math/rescue_prime_regular.rs
+++ b/twenty-first/src/shared_math/rescue_prime_regular.rs
@@ -801,7 +801,7 @@ pub struct RescuePrimeRegularState {
 
 impl RescuePrimeRegularState {
     #[inline]
-    fn new() -> RescuePrimeRegularState {
+    const fn new() -> RescuePrimeRegularState {
         RescuePrimeRegularState {
             state: [BFIELD_ZERO; STATE_SIZE],
         }