There's two patches here.

The first avoids recomputing constant field elements (#69 already has this, I just didn't see that PR).

The second patch tries different ways of handling overflows in the modular arithmetic, and picks a faster version for `add`. (One piece of the puzzle is that while the `if` looks scary, modern CPUs tend to have conditional move instructions, so the resulting code is actually branchless.)

Combining those two changes the code is faster than v0.8.0 on i7-6850K (so the Montgomery representation does indeed pay off):

```
rescue_prime_regular/hash_10/RescuePrimeRegular / Hash 10/10                                                                             
                        time:   [20.017 µs 20.045 µs 20.076 µs]
                        change: [-17.919% -17.781% -17.635%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

rescue_prime_regular/hash_varlen/RescuePrimeRegular / Hash Variable Length/16384                                                                           
                        time:   [35.703 ms 35.783 ms 35.877 ms]
                        change: [-14.785% -14.601% -14.401%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 50 measurements (12.00%)
  2 (4.00%) high mild
  4 (8.00%) high severe
```

Anyway, feel free to play around with this and cherry-pick whatever you like.