@int-e 2022-12-22
So why is the code faster now in this context? It looks like in the inner loop of `xlix_round()`, the compiler (llvm, I suppose) sees through the attempt to write branchless code inside `add` and produces a conditional jump instead!

```objdump
   b9a56:	b8 ff ff ff ff       	mov    $0xffffffff,%eax
   b9a5b:	48 29 da             	sub    %rbx,%rdx
   b9a5e:	72 02                	jb     b9a62 <_ZN12twenty_first11shared_math20rescue_prime_regular18RescuePrimeRegular10xlix_round17he700661c81f8c263E.llvm.5740233521524750972+0xa12>
   b9a60:	31 c0                	xor    %eax,%eax
   b9a62:
```

That's terrible. In contrast, the plain `if` results in a conditional move:

```objdump
   b942d:	48 29 da             	sub    %rbx,%rdx
   b9430:	b8 00 00 00 00       	mov    $0x0,%eax
   b9435:	49 0f 42 c5          	cmovb  %r13,%rax // %r13 is 0xFFFFFFFF here
```

No branch here, so the processor will be happy.