@int-e 2022-12-22
Interestingly this does not happen in isolation; compiling the following module

```rs
pub fn foo(a: u64, b: u64) -> u64 {
    let (r, c) = a.overflowing_sub(b);
    r.wrapping_sub(0u32.wrapping_sub(c as u32) as u64)
}

pub fn bar(a: u64, b: u64) -> u64 {
    let (r, c) = a.overflowing_sub(b);
    if c {
        r.wrapping_add(0xffffffff_00000001)
    } else {
        r
    }
}
```

results in identical code for both functions, using a conditional move:

```as
	movq	%rdi, %rcx
	subq	%rsi, %rcx
	movabsq	$-4294967295, %rax
	addq	%rcx, %rax
	subq	%rsi, %rdi
	cmovaeq	%rdi, %rax
	retq
```