@Sword-Smith 2022-12-22
Very, very nice. I can confirm that this gives a 100 % speedup (halving the running time) on the `rescue_prime_regular/hash_varlen/RescuePrimeRegular / Hash Variable Length/16384` and the a 60 % speedup for the `merkle_tree/merkle_tree/65536` benchmark.

Is this a compiler bug that we have uncovered? And if so, do you think it's in LLVM or in the `rustc`?

Thank you so much for uncovering this.