@jan-ferdinand 2022-12-22
I quite like the way to communicate which methods you have tried, and what the running times of those respective ways are.