@Sword-Smith 2022-12-23
Some snippets to perform this analysis. Full credit goes to @int-e 

```
> find target -name twenty_first\*
target/release/deps/twenty_first-60caca2773f2357d.d
target/release/deps/twenty_first-60caca2773f2357d
target/release/deps/twenty_first-242087d812946d75.d

> nm target/release/deps/twenty_first-60caca2773f2357d | grep xlix
00000000000b9050 t _ZN12twenty_first11shared_math20rescue_prime_regular18RescuePrimeRegular10xlix_round17he700661c81f8c263E.llvm.5740233521524750972

> objdump --disassemble=_ZN12twenty_first11shared_math20rescue_prime_regular18RescuePrimeRegular10xlix_round17he700661c81f8c263E.llvm.5740233521524750972 target/release/deps/twenty_first-60caca2773f2357d
[about 4000 lines of disassembly for twenty_first::shared_math::escue_prime_regular::RescuePrimeRegular::xlix_round()
```

http://paste.debian.net/plain/1265025

For isolated examples you can use
```
rustc -O --emit=asm --crate-type=lib
```